CREATED REGISTER TEACHER FUNCTIONALITY: Added register teacher (ActionResult) in AccountController  + moved Teacher-method (ActionResult) from CourseController to AccountController
