+++ MOAS-LMS/Views/Account/Teacher.cshtml
@@ -1,4 +1,4 @@
-﻿@*@model IEnumerable<MOAS_LMS.Models.TeacherModel>
+﻿@model IEnumerable<MOAS_LMS.Models.TeacherModel>
 
 @{
     ViewBag.Title = "Teacher";
@@ -43,4 +43,4 @@
         </tr>
     }
 
-</table>*@
+</table>
