+++ MOAS-LMS/Views/Shared/_Layout.cshtml
@@ -25,7 +25,7 @@
 
                     @if (User.IsInRole("Admin"))
                     {
-                    <li>@Html.ActionLink("Teachers", "Teacher", "Course")</li>
+                    <li>@Html.ActionLink("Teachers", "Teacher", "Account")</li>
                     }
                 </ul>
                 @Html.Partial("_LoginPartial")
